Added Twitter Library and basic tweet streaming
Added twitter streaming but now have more questions on scalability. Could tie this to mongoDB if my understanding is correct to filter down who's ID is who. Package I went with is Twitter-lite for those wanting to check it out. It seems better maintained that what I used to use.